Fix: Reorder the tests and update query logs function
